Followers
Create a Java program that keeps the information about Jane's Facebook followers, their likes, and comments. Keep a record of the followers, each with the count of likes and comments Jane has received from them.
You will be receiving lines with commands until you receive the "Log out" command. There are four possiblev commands:

1. "New follower: {username} ":
Add the username to your records (with O likes and O comments).
If a person with the given username already exists, ignore the line.
2. "Like: {username}: {count} ":
 If the username doesn't exist, add it to your records with the given count of likes.
 If the username exists, increase the count of likes with the given count.
3. "Comment: {username} ":
 If the username doesn't exist, add it to your records with 1 comment.
 If the username exists, increase the count of comments with 1.
4. "Blocked: {username}
Delete all records of the given username.
If it doesn't exist, print: "{Username} doesn't exist."

In the end, you have to print the count of followers, each follower with their likes and comments (the sum of likes and comments):
"{count} followers"
{username): {likes+comments}
{username}: {likes+comments}
...
{username}: {likes+comments}"